starting divide and conquer of older generations of code
